--- Server/src/Controller/getAllTrainingController.js
@@ -1,17 +1,0 @@
-const { training, companies } = require('../db');
-
-const getAll = async () => {
-    try {
-        const findAll = await training.findAll({
-            include: companies
-        })
-
-        return findAll
-    } catch (error) {
-        throw error
-    }
-}
-
-module.exports = {
-    getAll
-}
