--- Server/src/Controller/trainingDelete.js
@@ -1,14 +1,0 @@
-const { training } = require('../db')
-
-const delTraining = (id) => {
-    const trainingToDelete = training.destroy({
-        where:{
-            id: id
-        }
-    })
-    return trainingToDelete
-}
-
-module.exports = {
-    delTraining
-}
