+++ Server/src/Routes/index.js
@@ -2,15 +2,19 @@ const { Router } = require("express");
 const { trainingHandler } = require("../Handler/createTrainingHandler");
 const { getTrainingByName } = require("../Handler/getTrainingHandler");
 const { getAlltraining } = require("../Handler/getAllTraining");
-const {postOfferHandler,deleteOfferHandler,getOfferHandler,updateOfferHandler} = require("../Handler/offersHandler");
+const {
+  postOfferHandler,
+  deleteOfferHandler,
+  getOfferHandler,
+  updateOfferHandler,
+} = require("../Handler/offersHandler");
 const { loginHandler } = require("../Handler/loginHandler");
 
 const { deletetraining } = require('../Handler/deleteTraining');
 const { updateTraining } = require('../Handler/patchTraining')
 
-const { companiesSignUpHandler, getCompaniesHandler, updateCompanyHandler} = require("../Handler/companiesHandlers");
-const { userSignUpHandler, getUsersHandler, updateUserHandler} = require("../Handler/userHandlers");
-const { adminSignUpHandler, getAdminsHandler, updateAdminHandler} = require("../Handler/adminHandlers");
+const { deletetraining } = require("../Handler/deleteTraining");
+const { updateTraining } = require("../Handler/patchTraining");
 
 
 const router = Router();
@@ -45,13 +49,12 @@ router.get("/training", (req, res) => {
   }
 });
 
-router.post("/training/create",trainingHandler);
+router.post("/training/create", trainingHandler);
 
-router.delete('/training/delete', deletetraining)
-
+router.delete("/training/delete", deletetraining);
 
 //cambios
 
-router.patch("/training/update", updateTraining)
+router.patch("/training/update", updateTraining);
 
 module.exports = router;
